Disabled unexplained deferred shader loading behavior
CopyDev copies all of the shaders in the game folder to the platform folder for no explained reason.

All shaders load just fine from the game folder as far as I can tell. (I haven't tested with a game folder with an underscore in its name or other similar edge cases)